+++ Backend/index.js
@@ -4,6 +4,15 @@ const bodyParser = require('body-parser');
 const cors = require('cors');
 const http = require('http');
 const userRoutes = require("./routes/userRoutes");
+const spotifyRoutes = require('./routes/spotifyRoutes.js');
+const songRoutes = require('./routes/songRoutes.js');
+const playlistRoutes = require('./routes/playlistRoutes.js');
+const artistRoutes = require('./routes/artistRoutes.js');
+const recommendationRoutes = require('./routes/recommendationRoutes.js')
+const chatRoutes = require('./routes/chatRoutes.js')
+const messageRoutes = require('./routes/messageRoutes.js')
+const youtubeRoutes = require('./routes/ytMusicRoutes.js')
+const utilityRoutes = require('./routes/utilityRoutes.js')
 const { configureSocket } = require('./socket.js');
 
 require('./utils/cronJobMap.js');
@@ -34,6 +43,15 @@ app.get('/', (req, res) => {
 })
 
 app.use("/user", userRoutes);
+app.use("/spotify", spotifyRoutes);
+app.use("/songs", songRoutes);
+app.use("/playlist", playlistRoutes);
+app.use("/artists", artistRoutes);
+app.use("/recommendation", recommendationRoutes);
+app.use("/chat", chatRoutes);
+app.use("/message", messageRoutes);
+app.use("/yt", youtubeRoutes)
+app.use("/api", utilityRoutes)
 
 const PORT = process.env.PORT || 3000;
 server.listen(PORT, () => {
