+++ Backend/socket.js
@@ -11,10 +11,9 @@ function configureSocket(server) {
     });
 
     io.on('connection', (socket) => {
-        console.log('A user connected');
 
         socket.on('disconnect', () => {
-            console.log('User disconnected');
+
         });
     });
