+++ Backend/utils/tagValues.js
@@ -1,0 +1,55 @@
+const tagValues = {
+    genre: {
+        afrobeats: 1,
+        alternative: 2,
+        blues: 3,
+        classical: 4,
+        country: 5,
+        dance: 6,
+        disco: 7,
+        dubstep: 8,
+        electronic: 9,
+        folk: 10,
+        funk: 11,
+        gospel: 12,
+        grunge: 13,
+        hiphop: 14,
+        house: 15,
+        indie: 16,
+        jazz: 17,
+        metal: 18,
+        opera: 19,
+        pop: 20,
+        punk: 21,
+        rap: 22,
+        reggae: 23,
+        rock: 24,
+        rnb: 25,
+        soul: 26,
+        techno: 27,
+        trance: 28
+    },
+    mood: {
+        happy: 1,
+        sad: 2,
+        energetic: 3,
+        relaxed: 4,
+        angry: 5,
+        romantic: 6,
+        melancholic: 7,
+        upbeat: 8,
+        peaceful: 9,
+        moody: 10,
+        excited: 11,
+        nostalgic: 12
+    },
+    tempo: {
+        veryslow: 1,
+        slow: 2,
+        medium: 3,
+        fast: 4,
+        veryfast: 5
+    }
+};
+
+module.exports = tagValues;
