+++ src/components/errorboundary/ErrorBoundary.jsx
@@ -1,0 +1,32 @@
+import React, { Component } from 'react';
+
+class ErrorBoundary extends Component {
+    constructor(props) {
+        super(props);
+        this.state = { hasError: false };
+    }
+
+    static getDerivedStateFromError(error) {
+        return { hasError: true };
+    }
+
+    componentDidCatch(error, errorInfo) {
+        console.error("ErrorBoundary caught an error", error, errorInfo);
+    }
+
+    componentDidUpdate(prevProps, prevState) {
+        if (this.state.hasError && !prevState.hasError) {
+            window.location.reload();
+        }
+    }
+
+    render() {
+        if (this.state.hasError) {
+            return <h1>Something went wrong. Refreshing...</h1>;
+        }
+
+        return this.props.children;
+    }
+}
+
+export default ErrorBoundary;
