+++ src/components/requireauth/RequireAuth.jsx
@@ -4,7 +4,7 @@ import { Navigate } from 'react-router-dom';
 const RequireAuth = ({ children }) => {
     const JWT = localStorage.getItem('jwt');
     if (!JWT) {
-        return <Navigate to="/login" />;
+        return <Navigate to="/login" replace />;
     }
     return children;
 };
